Add allBill manualBill monthBill Component payMent
ยังไม่เสร็จ all manualBill monthBill
