+++ easyClassPay/app/pages/allBill.vue
@@ -1,0 +1,10 @@
+<template>
+    <div class="font-Bai-Jamjuree">
+        <navbar/>
+    </div>
+</template>
+
+<script setup>
+import navbar from '~/components/navbar.vue';
+import about from '~/components/about.vue';
+</script>
