+++ easyClassPay/app/pages/monthBill.vue
@@ -1,0 +1,14 @@
+<template>
+    <div class="font-Bai-Jamjuree">
+        <navbar/>
+
+
+    </div>
+</template>
+
+<script setup> 
+import navbar from '~/components/navbar.vue';
+import about from '~/components/about.vue';
+import payMent from '~/components/payMent.vue';
+
+</script>
